+++ packages/adhere-ui-comment/src/Comment/Info/index.tsx
@@ -3,6 +3,12 @@ import React from 'react';
 
 import Node, { defaultProps, propTypes } from '../../Components/Node';
 
+/**
+ * CommentInfo
+ * @param props
+ * @constructor
+ * @classdesc 评论节点
+ */
 function CommentInfo(props: INodeProps) {
   return (
     <Node {...props} isReply={false}>