+++ packages/adhere-ui-comment/src/Comment/List/index.tsx
@@ -21,6 +21,7 @@ class BackTopAnimationImpl extends BackTopAnimation {
 /**
  * CommentList
  * @constructor
+ * @classdesc 评论列表
  */
 function CommentList({
   className,