+++ packages/adhere-ui-comment/src/Reply/Info/index.tsx
@@ -3,6 +3,12 @@ import React from 'react';
 
 import Node, { defaultProps, propTypes } from '../../Components/Node';
 
+/**
+ * ReplyInfo
+ * @param props
+ * @constructor
+ * @classdesc 回复节点
+ */
 function ReplyInfo(props: INodeProps) {
   return <Node {...props} isReply />;
 }