+++ packages/adhere-ui-hooks/src/useSetState.tsx
@@ -1,12 +1,10 @@
 import { useState, useLayoutEffect, useRef } from 'react';
 
-/**
- * useSetState
- * @param defaultValue - 初始值
- */
-export default (defaultValue: any) => {
-  const [value, setValue] = useState<any>(defaultValue);
-  const callbackRef = useRef<() => {}>();
+type Dispatch<A> = (value: A, callback?: () => void) => void;
+
+function useSetState<T>(defaultValue: T): [T, Dispatch<T>] {
+  const [value, setValue] = useState<T>(defaultValue);
+  const callbackRef = useRef<Function>();
 
   useLayoutEffect(() => {
     callbackRef?.current?.();
@@ -14,9 +12,15 @@ export default (defaultValue: any) => {
 
   return [
     value,
-    (_value: any, callback: () => {}) => {
+    (_value, callback) => {
       callbackRef.current = callback;
       setValue(_value);
     },
   ];
-};
+}
+
+/**
+ * useSetState
+ * @param defaultValue - 初始值
+ */
+export default useSetState;