+++ packages/adhere-ui-searchtable/src/searchtable.tsx
@@ -141,6 +141,16 @@ abstract class SearchTable extends Suspense<ISearchTableProps, ISearchTableState
   abstract getOrderProp(): string;
 
   /**
+   * getOrderPropValue - 获取默认排序方式
+   */
+  abstract getOrderPropValue(): 'descend' | 'ascend';
+
+  /**
+   * getOrderFieldValue - 获取默认排序字段的值
+   */
+  abstract getOrderFieldValue(): string;
+
+  /**
    * onSubTableChange - 获取表格change句柄
    * @param pagination
    * @param filters
@@ -376,8 +386,8 @@ abstract class SearchTable extends Suspense<ISearchTableProps, ISearchTableState
     // @ts-ignore
     this.setState(
       {
-        [this.getOrderFieldProp()]: sorter.field,
-        [this.getOrderProp()]: sorter.order,
+        [this.getOrderFieldProp()]: sorter.field || this.getOrderFieldValue(),
+        [this.getOrderProp()]: sorter.order || this.getOrderPropValue(),
       },
       () => {
         const { order } = sorter;